requirements.txt
neo4j>=5.23
python-dotenv>=1.0
requests>=2.31
pydantic>=2.8
fastapi>=0.115
uvicorn>=0.30
langchain
langchain-community